Refactor some parts of IO
