+++ mne/epochs.py
@@ -19,7 +19,7 @@ from .io.write import (start_file, start_block, end_file, end_block,
                        write_int, write_float_matrix, write_float,
                        write_id, write_string)
 from .io.meas_info import read_meas_info, write_meas_info, _merge_info
-from .io.open import fiff_open
+from .io.open import fiff_open, _get_next_fname
 from .io.tree import dir_tree_find
 from .io.tag import read_tag
 from .io.constants import FIFF
@@ -2000,48 +2000,6 @@ def _is_good(e, ch_names, channel_type_idx, reject, flat, full_report=False,
             return True, None
         else:
             return False, bad_list
-
-
-def _get_next_fname(fid, fname, tree):
-
-    # Try to get the next filename tag for split files
-    nodes_list = dir_tree_find(tree, FIFF.FIFFB_REF)
-    next_fname = None
-    for nodes in nodes_list:
-        next_fname = None
-        for ent in nodes['directory']:
-            if ent.kind == FIFF.FIFF_REF_ROLE:
-                tag = read_tag(fid, ent.pos)
-                role = int(tag.data)
-                if role != FIFF.FIFFV_ROLE_NEXT_FILE:
-                    next_fname = None
-                    break
-            if ent.kind == FIFF.FIFF_REF_FILE_NAME:
-                tag = read_tag(fid, ent.pos)
-                next_fname = op.join(op.dirname(fname), tag.data)
-            if ent.kind == FIFF.FIFF_REF_FILE_NUM:
-                # Some files don't have the name, just the number. So
-                # we construct the name from the current name.
-                if next_fname is not None:
-                    continue
-                next_num = read_tag(fid, ent.pos).data
-                path, base = op.split(fname)
-                idx = base.find('.')
-                idx2 = base.rfind('-')
-                if idx2 < 0 and next_num == 1:
-                    # this is the first file, which may not be numbered
-                    next_fname = op.join(
-                        path, '%s-%d.%s' % (base[:idx], next_num,
-                                            base[idx + 1:]))
-                    continue
-                num_str = base[idx2 + 1:idx]
-                if not num_str.isdigit():
-                    continue
-                next_fname = op.join(path, '%s-%d.%s' % (base[:idx2],
-                                     next_num, base[idx + 1:]))
-        if next_fname is not None:
-            break
-    return next_fname
 
 
 @verbose
